+++ src/main/java/com/tiwilli/bechosystem/ClothesListController.java
@@ -4,8 +4,10 @@ import com.tiwilli.bechosystem.db.DbIntegrityException;
 import com.tiwilli.bechosystem.gui.listeners.DataChangeListener;
 import com.tiwilli.bechosystem.gui.util.Alerts;
 import com.tiwilli.bechosystem.gui.util.Utils;
-import com.tiwilli.bechosystem.model.entities.Category;
+import com.tiwilli.bechosystem.model.entities.Clothes;
 import com.tiwilli.bechosystem.model.services.CategoryService;
+import com.tiwilli.bechosystem.model.services.ClothesService;
+import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -22,24 +24,60 @@ import javafx.stage.Stage;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
+import java.util.Locale;
 import java.util.Optional;
 import java.util.ResourceBundle;
 
-public class CategoryListController implements Initializable, DataChangeListener {
+public class ClothesListController implements Initializable, DataChangeListener {
 
-    private CategoryService service;
+    private ClothesService service;
+
+    private CategoryService categoryService;
 
     @FXML
-    private TableView<Category> categoryTableView;
+    private TableView<Clothes> clothesTableView;
 
     @FXML
-    private TableColumn<Category, String> tableColumnName;
+    private TableColumn<Clothes, String> tableColumnName;
+
+    @FXML
+    private TableColumn<Clothes, String> tableColumnSize;
+
+    @FXML
+    private TableColumn<Clothes, String> tableColumnCategoryName;
+
+    @FXML
+    private TableColumn<Clothes, String> tableColumnStatus;
 
     @FXML
     private Label labelId;
 
     @FXML
     private Label labelName;
+
+    @FXML
+    private Label labelSize;
+
+    @FXML
+    private Label labelCategory;
+
+    @FXML
+    private Label labelPurchaseValue;
+
+    @FXML
+    private Label labelSalesValue;
+
+    @FXML
+    private Label labelPurchaseDate;
+
+    @FXML
+    private Label labelSalesDate;
+
+    @FXML
+    private Label labelPostDate;
+
+    @FXML
+    private Label labelStatus;
 
     @FXML
     private Button btRegister;
@@ -50,50 +88,59 @@ public class CategoryListController implements Initializable, DataChangeListener
     @FXML
     private Button btDelete;
 
-    private ObservableList<Category> observableList;
+    private ObservableList<Clothes> observableList;
 
-    public void setCategoryService(CategoryService service) {
+    public void setClothesService(ClothesService service) {
         this.service = service;
+    }
+
+    public void setCategoryService(CategoryService categoryService) {
+        this.categoryService = categoryService;
     }
 
     @FXML
     public void onBtRegisterAction(ActionEvent event) {
         Stage parentStage = Utils.currentStage(event);
-        Category obj = new Category();
-        createDialogForm(obj, "CategoryForm.fxml", parentStage);
+        Clothes obj = new Clothes();
+        createDialogForm(obj, "ClothesForm.fxml", parentStage);
     }
 
     @FXML
     public void onBtEditAction(ActionEvent event) {
-        Category obj = categoryTableView.getSelectionModel().getSelectedItem();
+        Clothes obj = clothesTableView.getSelectionModel().getSelectedItem();
         if (obj != null) {
             Stage parentStage = Utils.currentStage(event);
-            obj = service.findById(obj.getId());
-            createDialogForm(obj, "CategoryForm.fxml", parentStage);
-            categoryTableView.refresh();
+            createDialogForm(obj, "ClothesForm.fxml", parentStage);
+            clothesTableView.refresh();
         }
         else {
-            Alerts.showAlert("Nenhuma categoria selecionada", null, "Selecione uma categoria para editar", Alert.AlertType.WARNING);
+            Alerts.showAlert("Nenhuma peça selecionada", null, "Selecione uma peça para editar", Alert.AlertType.WARNING);
         }
     }
 
     @FXML
     public void onBtDeleteAction() {
-        Category obj = categoryTableView.getSelectionModel().getSelectedItem();
-        Optional<ButtonType> result = Alerts.showConfirmation("Confirmação", "Tem certeza que deseja excluir?");
+        Clothes obj = clothesTableView.getSelectionModel().getSelectedItem();
 
-        if (result.get() == ButtonType.OK) {
-            if (service == null) {
-                throw new IllegalStateException("Service was null");
+        if (obj != null) {
+            Optional<ButtonType> result = Alerts.showConfirmation("Confirmação", "Tem certeza que deseja excluir?");
+
+            if (result.get() == ButtonType.OK) {
+                if (service == null) {
+                    throw new IllegalStateException("Service was null");
+                }
+                try {
+                    service.remove(obj);
+                    updateTableView();
+                    clearItemsClothesTableView();
+                }
+                catch (DbIntegrityException e) {
+                    Alerts.showAlert("Erro!", null, "Você não pode remover uma categoria associada a um produto", Alert.AlertType.ERROR);
+                }
             }
-            try {
-                service.remove(obj);
-                updateTableView();
-                clearItemsCategoryTableView();
-            }
-            catch (DbIntegrityException e) {
-                Alerts.showAlert("Erro!", null, "Você não pode remover uma categoria associada a um produto", Alert.AlertType.ERROR);
-            }
+        }
+        else {
+            Alerts.showAlert("Nenhuma peça selecionada", null, "Selecione uma peça para excluir", Alert.AlertType.WARNING);
         }
     }
 
@@ -104,54 +151,77 @@ public class CategoryListController implements Initializable, DataChangeListener
 
     private void initializeNodes() {
         tableColumnName.setCellValueFactory(new PropertyValueFactory<>("name"));
+        tableColumnSize.setCellValueFactory(new PropertyValueFactory<>("size"));
+        showStatusDescription();
+        showCategoryName();
 
         Stage stage = (Stage) Main.getMainScene().getWindow();
-        categoryTableView.prefHeightProperty().bind(stage.heightProperty());
+        clothesTableView.prefHeightProperty().bind(stage.heightProperty());
 
-        categoryTableView.getSelectionModel().selectedItemProperty().addListener(
+        clothesTableView.getSelectionModel().selectedItemProperty().addListener(
                 (observable, oldValue, newValue) -> {
-                    selectItemsCategoryTableView(newValue);
+                    selectItemsClothesTableView(newValue);
                 }
         );
     }
 
-    private void selectItemsCategoryTableView(Category obj) {
+    private void selectItemsClothesTableView(Clothes obj) {
+
+        Locale.setDefault(Locale.US);
+
         if (obj != null) {
             labelId.setText(String.valueOf(obj.getId()));
             labelName.setText(obj.getName());
+            labelSize.setText(obj.getSize());
+            labelCategory.setText(obj.getCategory().getName());
+            labelPurchaseValue.setText(String.format("%.2f", obj.getPurchaseValue()));
+            labelSalesValue.setText(String.format("%.2f", obj.getSalesValue()));
+            labelPurchaseDate.setText(Utils.formatLabelDate(obj.getPurchaseDate(), "dd/MM/yyyy"));
+            labelSalesDate.setText(Utils.formatLabelDate(obj.getSalesDate(), "dd/MM/yyyy"));
+            labelPostDate.setText(Utils.formatLabelDate(obj.getPostDate(), "dd/MM/yyyy"));
+            labelStatus.setText(obj.getStatus().getDescription());
         }
         else {
             return;
         }
     }
 
-    private void clearItemsCategoryTableView() {
+    private void clearItemsClothesTableView() {
         labelId.setText("");
         labelName.setText("");
+        labelSize.setText("");
+        labelPurchaseValue.setText("");
+        labelSalesValue.setText("");
+        labelPurchaseDate.setText("");
+        labelSalesDate.setText("");
+        labelPostDate.setText("");
+        labelStatus.setText("");
+        labelCategory.setText("");
     }
 
     public void updateTableView() {
         if (service == null) {
             throw new IllegalStateException("Service was null");
         }
-        List<Category> list = service.findAll();
+        List<Clothes> list = service.findAll();
         observableList = FXCollections.observableArrayList(list);
-        categoryTableView.setItems(observableList);
+        clothesTableView.setItems(observableList);
     }
 
-    private void createDialogForm(Category obj, String absoluteName, Stage parentStage) {
+    private void createDialogForm(Clothes obj, String absoluteName, Stage parentStage) {
             try {
                 FXMLLoader loader = new FXMLLoader(getClass().getResource(absoluteName));
                 Pane pane = loader.load();
 
-                CategoryFormController controller = loader.getController();
-                controller.setCategory(obj);
-                controller.setCategoryService(new CategoryService());
+                ClothesFormController controller = loader.getController();
+                controller.setClothes(obj);
+                controller.setServices(new ClothesService(), new CategoryService());
+                controller.loadAssociatedObjects();
                 controller.subscribeDataChangeListener(this);
                 controller.updateFormData();
 
                 Stage dialogStage = new Stage();
-                dialogStage.setTitle("Entre com os dados da categoria");
+                dialogStage.setTitle("Entre com os dados da peça");
                 dialogStage.setScene(new Scene(pane));
                 dialogStage.setResizable(false);
                 dialogStage.initOwner(parentStage);
@@ -167,5 +237,39 @@ public class CategoryListController implements Initializable, DataChangeListener
     @Override
     public void onDataChanged() {
         updateTableView();
+    }
+
+    private void showStatusDescription() {
+        tableColumnStatus.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getStatus().getDescription()));
+        tableColumnStatus.setCellFactory(column -> new TableCell<Clothes, String>() {
+            @Override
+            protected void updateItem(String statusDescription, boolean empty) {
+                super.updateItem(statusDescription, empty);
+
+                if (statusDescription == null) {
+                    setText(null);
+                }
+                else {
+                    setText(statusDescription);
+                }
+            }
+        });
+    }
+
+    private void showCategoryName() {
+        tableColumnCategoryName.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getCategory().getName()));
+        tableColumnCategoryName.setCellFactory(column -> new TableCell<Clothes, String>() {
+            @Override
+            protected void updateItem(String categoryName, boolean empty) {
+                super.updateItem(categoryName, empty);
+
+                if (categoryName == null) {
+                    setText(null);
+                }
+                else {
+                    setText(categoryName);
+                }
+            }
+        });
     }
 }
